requirements.txt
# Core API Clients
deepgram-sdk>=3.0.0
openai>=1.0.0
groq>=0.4.0

# Web Server
fastapi
uvicorn
requests
websockets

# Audio Processing (Lightweight)
numpy
scipy

# Configuration
python-dotenv